+++ doc/source/changelog.rst
@@ -12,6 +12,13 @@ Development was migrated into the
 `OpenBioSim <https://github.com/openbiosim>`__
 organisation on `GitHub <https://github.com/openbiosim/sire>`__.
 
+`2025.3.0 <https://github.com/openbiosim/sire/compare/2025.2.0...2025.3.0>`__ - December 2025
+---------------------------------------------------------------------------------------------
+
+* Reset GCMC water state prior to minimisation following a dynamics crash.
+
+* Set appropriate default OpenMM box vectors when a system has no periodic space.
+
 `2025.2.0 <https://github.com/openbiosim/sire/compare/2025.1.0...2025.2.0>`__ - October 2025
 --------------------------------------------------------------------------------------------